machin

+++ monitor.h
@@ -2,6 +2,6 @@
 #define MONITOR_H
 
 void init_monitor();
-double get_current_angle();
+float get_current_angle();
 
 #endif
